[FLINK-10600] Provide End-to-end test cases for modern Kafka connectors
