+++ flink-connectors/flink-connector-kafka/pom.xml
@@ -52,6 +52,10 @@ under the License.
 					<groupId>org.apache.kafka</groupId>
 					<artifactId>kafka_${scala.binary.version}</artifactId>
 				</exclusion>
+				<exclusion>
+					<groupId>org.apache.kafka</groupId>
+					<artifactId>kafka-clients</artifactId>
+				</exclusion>
 			</exclusions>
 		</dependency>
