+++ flink-end-to-end-tests/test-scripts/test_streaming_kafka.sh
@@ -17,8 +17,11 @@
 # limitations under the License.
 ################################################################################
 
+set -e
+set -u
+set -o pipefail
+
 source "$(dirname "$0")"/common.sh
-source "$(dirname "$0")"/kafka-common.sh 0.10.2.0 3.2.0 3.2
+source "$(dirname "$0")"/kafka-common.sh 2.0.0 5.0.0 5.0
 
-source "$(dirname "$0")"/test_streaming_kafka_common.sh $FLINK_DIR/examples/streaming/Kafka010Example.jar
-
+source "$(dirname "$0")"/test_streaming_kafka_common.sh $FLINK_DIR/examples/streaming/KafkaExample.jar
